Use the defaultChannel for the OADP OLM Subscription
Since a default channel is specified in the package manifest of the OADP
operator OLM bundle, we can use it to get the latest stable version of
the latter. This way we won't have to update the Subscription whenever a
new version comes up or a channel we are currently using suddencly
disappears.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>