Set empty default IBI release registry
When setting this default to quay.io and the seed image is generated
with a CI release registry, i.e. registry.ci.openshift.org, the IBI SNO
reconfiguration ends up with a non-existent cluster-version-operator image
pull-spec. The result is an SNO with a non-running CVO, thus the
installation never succeeds (even if all other components are healthy).

Signed-off-by: Michail Resvanis <mresvani@redhat.com>